refactor(service): reorganization of service package
